% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset.label}
\alias{asset.label}
\title{Data for clustering and labeling ETFs}
\format{
A data frame with 44 observations (rows) on 3 variables (columns)
\describe{
\item{id}{name of ETF}
\item{label}{from 1 to 5 according to the specific asset class}
\item{class}{specific asset class (5 categories)}
}
}
\source{
{Created in-house to serve as an example}
}
\usage{
asset.label
}
\description{
Contains asset class of ETFs
}
\examples{
data(asset.label)
}
\keyword{datasets}
