% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{create_overlapping_windows}
\alias{create_overlapping_windows}
\title{Function to create overlapping windows.}
\usage{
create_overlapping_windows(data, FT = 1290, ov = 125)
}
\arguments{
\item{data}{vector or data frame}

\item{FT}{Window size. By default set to 1290 training days (five years).}

\item{ov}{Number of different days for two consecutive time windows.. By default set to 125 training days (six months).}
}
\value{
a list containing the rolling windows
}
\description{
Function to create overlapping windows.
}
