% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{g_ret}
\alias{g_ret}
\title{Function to compute gross returns.}
\usage{
g_ret(x)
}
\arguments{
\item{x}{data object containing ordered price observations}
}
\value{
An object of the same class as x with the gross returns
}
\description{
Calculate gross returns from prices.
}
