% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{plot_LR_cp}
\alias{plot_LR_cp}
\title{Plot methods for a LR_cp object}
\usage{
plot_LR_cp(LR_cp_measure, asset_label)
}
\arguments{
\item{LR_cp_measure}{object of class LR_cp}

\item{asset_label}{vector containing asset label}
}
\value{
plot of LR_cp measures by asset classes
}
\description{
Plot methods for a LR_cp object
}
