% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{plot_ptf}
\alias{plot_ptf}
\title{Plot methods for a ptf_construction object}
\usage{
plot_ptf(ptf.oos.values)
}
\arguments{
\item{ptf.oos.values}{object of class ptf_construction}
}
\value{
plot oos portfolio values
}
\description{
Plot methods for a ptf_construction object
}
