% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.data}
\alias{sample.data}
\title{Data for infoset function}
\format{
A data frame with 3174 rows and 44 columns
}
\source{
{Created in-house to serve as an example}
}
\usage{
sample.data
}
\description{
Contains daily prices of ETFs
}
\examples{
data(sample.data)
}
\keyword{datasets}
