% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.data.ts}
\alias{sample.data.ts}
\title{Data with time points for portfolio construction using the LR_cp measure}
\format{
A data frame with 3175 rows and 45 columns
}
\source{
{Created in-house to serve as an example}
}
\usage{
sample.data.ts
}
\description{
Contains daily prices of ETFs
}
\examples{
data(sample.data.ts)
}
\keyword{datasets}
