% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{summary_ptf}
\alias{summary_ptf}
\title{Plot methods for a ptf_construction object}
\usage{
summary_ptf(ptf.oos.values)
}
\arguments{
\item{ptf.oos.values}{object of class ptf_construction}
}
\value{
summary of oos portfolio values
}
\description{
Plot methods for a ptf_construction object
}
