% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{tail_mixture}
\alias{tail_mixture}
\title{Function to find the most-left distribution set.}
\usage{
tail_mixture(y, shift, n_it, plot)
}
\arguments{
\item{y}{vector or data frame}

\item{shift}{double}

\item{n_it}{integer}

\item{plot}{option}
}
\value{
data object
}
\description{
An adaptive clustering algorithm identifies sub-groups of gross returns at each iteration by approximating their distribution with a
sequence of two-component log-normal mixtures.
}
